+++ src/main/java/de/fraunhofer/isst/dataspaceconnector/controller/ResourceDataController.java
@@ -4,9 +4,10 @@ import de.fraunhofer.isst.dataspaceconnector.exceptions.contract.ContractExcepti
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.InvalidResourceException;
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.ResourceException;
 import de.fraunhofer.isst.dataspaceconnector.exceptions.resource.ResourceNotFoundException;
+import de.fraunhofer.isst.dataspaceconnector.model.QueryInput;
 import de.fraunhofer.isst.dataspaceconnector.services.resources.OfferedResourceServiceImpl;
 import de.fraunhofer.isst.dataspaceconnector.services.resources.RequestedResourceServiceImpl;
-import de.fraunhofer.isst.dataspaceconnector.services.resources.ResourceService;
+import de.fraunhofer.isst.dataspaceconnector.services.utils.ValidationUtils;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
@@ -17,7 +18,13 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.bind.annotation.RestController;
 
 import java.util.UUID;
 
@@ -33,18 +40,20 @@ public class ResourceDataController { // Header: Content-Type: application/json
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ResourceDataController.class);
 
-    private final ResourceService offeredResourceService, requestedResourceService;
+    private final OfferedResourceServiceImpl offeredResourceService;
+    private final RequestedResourceServiceImpl requestedResourceService;
 
     /**
      * Constructor for ResourceDataController.
      *
-     * @param offeredResourceService The service for the offered resources
-     * @param requestedResourceService The service for the requested resources
+     * @param offeredResourceService The service for the offered resources.
+     * @param requestedResourceService The service for the requested resources.
      * @throws IllegalArgumentException if any of the parameters is null.
      */
     @Autowired
     public ResourceDataController(OfferedResourceServiceImpl offeredResourceService,
-        RequestedResourceServiceImpl requestedResourceService) throws IllegalArgumentException {
+                                  RequestedResourceServiceImpl requestedResourceService)
+            throws IllegalArgumentException {
         if (offeredResourceService == null)
             throw new IllegalArgumentException("The OfferedResourceService cannot be null.");
 
@@ -76,7 +85,7 @@ public class ResourceDataController { // Header: Content-Type: application/json
             example = "a4212311-86e4-40b3-ace3-ef29cd687cf9")
         @PathVariable("resource-id") UUID id,
         @Parameter(description = "The resource data.", required = true, example = "Data String")
-        @RequestParam("data") String data) {
+        @RequestBody String data) {
         try {
             offeredResourceService.addData(id, data);
             return new ResponseEntity<>("", HttpStatus.CREATED);
@@ -107,15 +116,20 @@ public class ResourceDataController { // Header: Content-Type: application/json
             @ApiResponse(responseCode = "200", description = "Ok"),
             @ApiResponse(responseCode = "404", description = "Not found"),
             @ApiResponse(responseCode = "500", description = "Internal server error")})
-    @RequestMapping(value = "/{resource-id}/data", method = RequestMethod.GET)
+    @RequestMapping(value = "/{resource-id}/data", method = RequestMethod.POST)
     // params = {"type=string"} NOT SUPPORTED with OpenAPI
     @ResponseBody
     public ResponseEntity<String> getDataById(@Parameter(description = "The resource uuid.",
             required = true, example = "a4212311-86e4-40b3-ace3-ef29cd687cf9")
-                                              @PathVariable("resource-id") UUID id) {
+            @PathVariable("resource-id") UUID id,
+            @Parameter(description = "The query parameters and headers to use when fetching the " +
+                      "data from the backend system.")
+              @RequestBody(required = false) QueryInput queryInput) {
         try {
+            ValidationUtils.validateQueryInput(queryInput);
             try {
-                return new ResponseEntity<>(offeredResourceService.getData(id), HttpStatus.OK);
+                return new ResponseEntity<>(offeredResourceService.getData(id, queryInput),
+                        HttpStatus.OK);
             } catch (ResourceNotFoundException offeredResourceServiceException) {
                 LOGGER.debug(
                         "Could not find resource in OfferedResourceService. [id=({}), exception=({})]",
@@ -141,6 +155,16 @@ public class ResourceDataController { // Header: Content-Type: application/json
                     id, exception.getMessage());
             return new ResponseEntity<>("The deposited policy cannot be enforced.",
                     HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (ResourceException exception) {
+            LOGGER.warn("Failed to retrieve data. [id=({}), exception=({})]", id,
+                    exception.getMessage());
+            return new ResponseEntity<>("Failed to load data.",
+                    HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (IllegalArgumentException exception) {
+            LOGGER.debug("Failed to fetch resource data. [id=({}), exception=({})]",
+                    id, exception.getMessage());
+            return new ResponseEntity<>("Invalid query input.",
+                    HttpStatus.BAD_REQUEST);
         } catch (Exception exception) {
             LOGGER.warn("Failed to load resource. [id=({}), exception=({})]", id,
                     exception.getMessage());
@@ -162,18 +186,22 @@ public class ResourceDataController { // Header: Content-Type: application/json
             @ApiResponse(responseCode = "200", description = "Ok"),
             @ApiResponse(responseCode = "404", description = "Not found"),
             @ApiResponse(responseCode = "500", description = "Internal server error")})
-    @RequestMapping(value = "/{resource-id}/{representation-id}/data", method = RequestMethod.GET)
+    @RequestMapping(value = "/{resource-id}/{representation-id}/data", method = RequestMethod.POST)
     @ResponseBody
     public ResponseEntity<String> getDataByRepresentation(
             @Parameter(description = "The resource uuid.", required = true,
                     example = "a4212311-86e4-40b3-ace3-ef29cd687cf9")
             @PathVariable("resource-id") UUID resourceId,
             @Parameter(description = "The representation uuid.", required = true)
-            @PathVariable("representation-id") UUID representationId) {
+            @PathVariable("representation-id") UUID representationId,
+            @Parameter(description = "The query parameters and headers to use when fetching the " +
+                    "data from the backend system.")
+            @RequestBody(required = false) QueryInput queryInput) {
         try {
+            ValidationUtils.validateQueryInput(queryInput);
             try {
                 return new ResponseEntity<>(
-                        offeredResourceService.getDataByRepresentation(resourceId, representationId, null),
+                        offeredResourceService.getDataByRepresentation(resourceId, representationId, queryInput),
                         HttpStatus.OK);
             } catch (ResourceNotFoundException offeredResourceServiceException) {
                 LOGGER.debug(
@@ -200,6 +228,16 @@ public class ResourceDataController { // Header: Content-Type: application/json
                     resourceId, exception.getMessage());
             return new ResponseEntity<>("The deposited policy cannot be enforced.",
                     HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (ResourceException exception) {
+            LOGGER.warn("Failed to retrieve data. [id=({}), exception=({})]", resourceId,
+                    exception.getMessage());
+            return new ResponseEntity<>("Failed to load data.",
+                    HttpStatus.INTERNAL_SERVER_ERROR);
+        } catch (IllegalArgumentException exception) {
+            LOGGER.debug("Failed to fetch resource data. [id=({}), exception=({})]",
+                    resourceId, exception.getMessage());
+            return new ResponseEntity<>("Invalid query input.",
+                    HttpStatus.BAD_REQUEST);
         } catch (Exception exception) {
             LOGGER.warn("Failed to load resource. [id=({}), exception=({})]", resourceId,
                     exception.getMessage());