+++ src/main/java/de/fraunhofer/isst/dataspaceconnector/services/messages/handler/ArtifactRequestHandler.java
@@ -171,7 +171,7 @@ public class ArtifactRequestHandler implements MessageHandler<ArtifactRequestMes
 
                 try {
                     // Check if the policy allows data access. TODO: Change to contract agreement. (later)
-                    if (policyHandler.onDataProvision(resourceMetadata.getPolicy())) {
+                    if (policyHandler.onDataProvision(resourceMetadata.getPolicy(), requestMessage.getIssuerConnector())) {
                         String data;
 
                         try {
@@ -204,6 +204,16 @@ public class ArtifactRequestHandler implements MessageHandler<ArtifactRequestMes
                                 .withDefaultHeader(RejectionReason.INTERNAL_RECIPIENT_ERROR,
                                     "Something went wrong.", connector.getId(),
                                     connector.getOutboundModelVersion());
+                        } catch (IllegalArgumentException exception) {
+                            LOGGER.warn("Failed to fetch resource data. [id=({}), " +
+                                            "resourceId=({}), artifactId=({}), exception=({})]",
+                                    requestMessage.getId(), resourceId, artifactId,
+                                    exception.getMessage());
+                            return ErrorResponse
+                                    .withDefaultHeader(RejectionReason.BAD_PARAMETERS,
+                                            "Invalid query input.",
+                                            connector.getId(),
+                                            connector.getOutboundModelVersion());
                         }
 
                         // Build artifact response.