+++ src/main/java/de/fraunhofer/isst/dataspaceconnector/services/messages/implementation/NotificationMessageService.java
@@ -7,7 +7,6 @@ import de.fraunhofer.iais.eis.util.Util;
 import de.fraunhofer.isst.dataspaceconnector.exceptions.message.MessageBuilderException;
 import de.fraunhofer.isst.dataspaceconnector.services.messages.MessageService;
 import de.fraunhofer.isst.dataspaceconnector.services.resources.OfferedResourceServiceImpl;
-import de.fraunhofer.isst.dataspaceconnector.services.utils.IdsUtils;
 import de.fraunhofer.isst.ids.framework.communication.http.IDSHttpService;
 import de.fraunhofer.isst.ids.framework.configuration.ConfigurationContainer;
 import de.fraunhofer.isst.ids.framework.configuration.SerializerProvider;
@@ -36,18 +35,14 @@ public class NotificationMessageService extends MessageService {
      * @param idsHttpService The service for ids messaging
      * @param configurationContainer The container with the configuration
      * @param resourceService The service for resources
-     * @param idsUtils The utilities for ids messages
      * @param serializerProvider The service for serializing
      * @throws IllegalArgumentException if any of the parameters is null
      */
     @Autowired
     public NotificationMessageService(DapsTokenProvider tokenProvider, IDSHttpService idsHttpService,
         ConfigurationContainer configurationContainer, OfferedResourceServiceImpl resourceService,
-        IdsUtils idsUtils, SerializerProvider serializerProvider) throws IllegalArgumentException {
-        super(idsHttpService, idsUtils, serializerProvider, resourceService);
-
-        if (configurationContainer == null)
-            throw new IllegalArgumentException("The ConfigurationContainer cannot be null.");
+        SerializerProvider serializerProvider) throws IllegalArgumentException {
+        super(idsHttpService, serializerProvider, resourceService, configurationContainer);
 
         if (tokenProvider == null)
             throw new IllegalArgumentException("The TokenProvider cannot be null.");